fixes for device/save, command/update

+++ src/main/java/com/devicehive/service/DeviceService.java
@@ -1,6 +1,5 @@
 package com.devicehive.service;
 
-import com.devicehive.configuration.Constants;
 import com.devicehive.dao.*;
 import com.devicehive.exceptions.HiveException;
 import com.devicehive.model.*;
@@ -17,8 +16,6 @@ import org.slf4j.LoggerFactory;
 import javax.ejb.Stateless;
 import javax.inject.Inject;
 import javax.interceptor.Interceptors;
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
 import javax.websocket.Session;
 import java.util.Date;
 import java.util.List;
@@ -28,8 +25,6 @@ import java.util.Set;
 @Stateless
 public class DeviceService {
     private static final Logger logger = LoggerFactory.getLogger(DeviceService.class);
-    @PersistenceContext(unitName = Constants.PERSISTENCE_UNIT)
-    private EntityManager em;
     @Inject
     private DeviceCommandDAO deviceCommandDAO;
     @Inject
@@ -97,7 +92,7 @@ public class DeviceService {
                                                            DeviceEquipment deviceEquipment) {
         if (deviceEquipment != null) {
             if (deviceEquipmentDAO.update(deviceEquipment) == 0) {
-                logger.warn("No equipments to update found");
+                deviceEquipmentDAO.saveDeviceEquipment(deviceEquipment);
             }
         }
         notification.setDevice(device);
@@ -119,7 +114,7 @@ public class DeviceService {
     public DeviceClass createOrUpdateDeviceClass(DeviceClass deviceClass, Set<Equipment> newEquipmentSet) {
         DeviceClass stored;
         if (deviceClass.getId() != null) {
-            stored = em.find(DeviceClass.class, deviceClass.getId());
+            stored = deviceClassDAO.getDeviceClass(deviceClass.getId());
         } else {
             stored = deviceClassDAO.getDeviceClassByNameAndVersion(deviceClass.getName(),
                     deviceClass.getVersion());
@@ -135,7 +130,7 @@ public class DeviceService {
             return stored;
         } else {
             //create
-            em.persist(deviceClass);
+            deviceClassDAO.saveDeviceClass(deviceClass);
             updateEquipment(newEquipmentSet, deviceClass);
             return deviceClass;
         }
@@ -144,28 +139,19 @@ public class DeviceService {
 
     public void updateEquipment(Set<Equipment> newEquipmentSet, DeviceClass deviceClass) {
         List<Equipment> existingEquipments = equipmentDAO.getByDeviceClass(deviceClass);
-        if (!newEquipmentSet.isEmpty()) {
-            if (!existingEquipments.isEmpty()) {
+        if (!newEquipmentSet.isEmpty() && !existingEquipments.isEmpty()) {
                 equipmentDAO.removeEquipment(existingEquipments);
-            } else {
-                for (Equipment equipment : newEquipmentSet) {
-                    DeviceEquipment deviceEquipment = new DeviceEquipment();
-                    deviceEquipment.setTimestamp(new Date(0));
-                    deviceEquipment.setCode(equipment.getCode());
-                    em.persist(deviceEquipment);
-                }
-            }
         }
         for (Equipment equipment : newEquipmentSet) {
             equipment.setDeviceClass(deviceClass);
-            em.persist(equipment);
+            equipmentDAO.saveEquipment(equipment);
         }
     }
 
     public void createOrUpdateDevice(Device device) {
-        Device existingDevice = deviceDAO.findByUUIDForWrite(device.getGuid());
+        Device existingDevice = deviceDAO.findByUUID(device.getGuid());
         if (existingDevice == null) {
-            em.persist(device);
+            deviceDAO.saveDevice(device);
         } else {
             existingDevice.setDeviceClass(device.getDeviceClass());
             existingDevice.setStatus(device.getStatus());
